Fix share by email password validation. Refactor share context provider.
