+++ packages/react-mui-store-file-viewer/src/components/CreateNewShareButton/CreateSharesButton/ShareAuthenticatorField.tsx
@@ -41,7 +41,8 @@ const UnControlledShareAuthenticatorField: FC<
 > = ({ onValueChange, ...props }) => {
   const value: AuthenticatorValue = props.value;
   const theme = useTheme();
-  const { contentType, password, loading } = useCreateShareContext();
+  const { shareType, contentType, password, loading, setShareLink } =
+    useCreateShareContext();
   const [authenticatorType, setAuthenticatorType] =
     useState<ObjectStore.ShareAuthenticatorType>(
       ObjectStore.ShareAuthenticatorType.Sms
@@ -58,6 +59,18 @@ const UnControlledShareAuthenticatorField: FC<
     newType: string
   ) => {
     if (!newType) return;
+    if (onValueChange && shareType === "email") {
+      onValueChange({
+        ...value,
+        authenticators: value.authenticators.map((r) => {
+          return {
+            ...r,
+            auth_type: newType,
+          };
+        }),
+      });
+    }
+    setShareLink(undefined);
     setAuthenticatorType(newType as ObjectStore.ShareAuthenticatorType);
   };
 
@@ -68,7 +81,7 @@ const UnControlledShareAuthenticatorField: FC<
 
   const handleRecipientChange = (recipients: ObjectStore.Recipient[]) => {
     if (!onValueChange) return;
-    const authType = value.authenticatorType;
+    const authType = authenticatorType;
 
     onValueChange({
       ...value,
@@ -111,10 +124,6 @@ const UnControlledShareAuthenticatorField: FC<
         },
       ],
     });
-  };
-
-  const handleReset = (authType: string) => {
-    handleAuthenticatorTypeChange(undefined, authType);
   };
 
   return (
@@ -256,10 +265,7 @@ const UnControlledShareAuthenticatorField: FC<
                   Don’t forget to copy this password to your clipboard.
                 </Typography>
               </Stack>
-              <GeneratePasswordField
-                //value={password}
-                onValueChange={handlePasswordChange}
-              />
+              <GeneratePasswordField onValueChange={handlePasswordChange} />
             </>
           )}
         </Stack>
